Create 12_Mersenne Number.c
A Mersenne Number is a positive integer that is one less than a power of two. For example, the n-
th Mersenne number, Mn. is: Mn = 2 n -l
A prime number is a positive integer that has exactly two distinct positive integer divisors, namely
1 and itself. The number 1 is, by definition, not a prime number. A Mersenne prime is a Mersenne
number that is prime. The number 3 is the smallest Mersenne prime; because 3 is a prime number,
and 2 2 -1=3. As of February 2014, 48 Mersenne primes are known. Write a complete C program
that finds and prints the 5 smallest Mersenne prime, each by itself on a line. You may use any C
math library functions in your solution.